+++ src/sticky_note/sticky-note-default.ts
@@ -1,6 +1,6 @@
-import deleteButton from '../action_buttons/delete';
-import editButton from '../action_buttons/edit';
-import renameButton from '../action_buttons/rename';
+import deleteButton from '../actions/delete';
+import editButton from '../actions/edit';
+import renameButton from '../actions/rename';
 import MeasureOfLength from '../value_objects/measue-of-length';
 import Title from '../value_objects/title';
 import StickyNote from './sticky-note';
